pyproject.toml
[tool.poetry]
name = "mylms"
version = "0.1.0"
description = ""
authors = ["Tatyana Kavardaeva <tatyana.kavardaeva@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.3"
python-dotenv = "1.0.1"
djangorestframework = "3.15.2"
psycopg2-binary = "2.9.10"
pillow = "11.0.0"
ipython = "8.29.0"
djangorestframework-simplejwt = "5.3.1"
django = "5.1.3"
coverage = "7.6.8"
drf-yasg = "1.21.8"
django-cors-headers = "4.6.0"
celery = "5.4.0"
eventlet = "0.38.0"
django-celery-beat = "2.7.0"
redis = "5.2.0"
flake8 = "^7.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
